% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{AllIntegers}
\alias{AllIntegers}
\title{Test if all items in a variable are integers}
\usage{
AllIntegers(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
logical.
}
\description{
Test if all items in a variable are integers
}
