% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{AllVariablesNames}
\alias{AllVariablesNames}
\title{Find the names of the variables in a formula}
\usage{
AllVariablesNames(formula, data = NULL)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}} from which to extract variable names if \code{.}
is used in the formula.}
}
\description{
Handles \code{.} on right hand side of formula and \code{$} within backticks in
variable names.
}
