% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{AnyNA}
\alias{AnyNA}
\title{\code{AnyNA}}
\usage{
AnyNA(data, formula = NULL)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{formula}{A \code{\link{formula}}. Where supplied, only variables in the formula are checked.}
}
\value{
logical.
}
\description{
Checks to see if there are any NAs in a data frame.0.
}
