% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{AnyNegative}
\alias{AnyNegative}
\title{#' \code{UnclassIfNecessary}
#' @description Unclasses a variable if it is a factor. Otherwise, returns x.
#' @param x A vector.
#' @return A vector
#' @export
UnclassIfNecessary <- function(x)
{
    if(is.factor(x))
        return(unclass(x));
    return(x);
}
\code{AnyNegative}}
\usage{
AnyNegative(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
logical.
}
\description{
The values contain a negative value.
}
