% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{CheckVariableLengths}
\alias{CheckVariableLengths}
\title{CheckVariableLengths}
\usage{
CheckVariableLengths(variable.list, list.name)
}
\arguments{
\item{variable.list}{A list of variables from a form.dropBox control with
multiple inputs.}

\item{list.name}{The name used to refer to the variable list in the error
message.}
}
\value{
Nothing is returned, but an error is thrown if the variables have
    differing lengths.
}
\description{
CheckVariableLengths
}
