% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{CollectWarnings}
\alias{CollectWarnings}
\title{Collect warnings and just warn once.}
\usage{
CollectWarnings(expr, return.list = FALSE)
}
\arguments{
\item{expr}{R expression}

\item{return.list}{When \code{TRUE} return a list with
list(object,warnings) instead of issuing the warnings.
Otherwise, when \code{FALSE} issue the warnings and return the
object.}
}
\value{
The value of the expression or a list with the value of
    the expression and a list of warning messages
}
\description{
Collect warnings and just warn once.
}
