% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{DataTableWithRItemFormat}
\alias{DataTableWithRItemFormat}
\title{\code{DataTableWithRItemFormat}}
\usage{
DataTableWithRItemFormat(dd, caption = NULL, header.alignments = NULL)
}
\arguments{
\item{dd}{The \code{data frame} that you want to show as a widget.}

\item{caption}{Text to be placed beneath the table.}

\item{header.alignments}{A character vector with one element for each column in 
\code{dd} which specifies whether the cells of each column should be left or right
aligned. Acceptable values are \code{"right"} and \code{"left"}. If you don't 
specify the alignments then columns that look like text will be left-aligned, and
columns that look like numbers will be right-aligned.}
}
\description{
Create an HTML widget DataTable which looks like Q's blue table style.
Facilitates nicer printing of wrapping text and coefficient matrices.
}
\examples{
my.df <- data.frame(First = c(1,2,3), Second = c("a", "b", "c"))
my.dt <- DataTableWithRItemFormat(my.df, caption = "A nice table")
my.dt
}

