% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{DIM}
\alias{DIM}
\title{Determine the dimensions of an array or vector}
\usage{
DIM(x)
}
\arguments{
\item{x}{The input to be inspected and have its dimensions returned}
}
\value{
An integer vector containing either the length of the vector
or the integer dimensions of the array.
}
\description{
Used to give the dimensions of any array or vector.
A vector is considered 1 dimensional and this function will return its
length.
}
\seealso{
Extension for \code{dim} in the same vein as \code{ncol} to \code{NCOL}.
}
