% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{EscapeRegexSymbols}
\alias{EscapeRegexSymbols}
\title{Escape any special regex characters in a character vector}
\usage{
EscapeRegexSymbols(text)
}
\arguments{
\item{text}{A character vector}
}
\description{
Used to escape special characters from text before passing it
  to regex functions such as gsub.
}
