% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{ExpectWarning}
\alias{ExpectWarning}
\title{\code{ExpectWarning}}
\usage{
ExpectWarning(code, string)
}
\arguments{
\item{code}{The code to evaluate for warnings.}

\item{string}{The string of text to be searched for in the warnings}
}
\description{
Checks to see if the printing of an object causes warnings to appear.
}
