% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{getIsStringFromControl}
\alias{getIsStringFromControl}
\title{Check whether a character vector was generated by a Displayr control.
noRd}
\usage{
getIsStringFromControl(x)
}
\description{
Check whether a character vector was generated by a Displayr control.
noRd
}
