% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{GetTranslatedQuestionType}
\alias{GetTranslatedQuestionType}
\title{Get the Q/Displayr name for an entry in a questiontype attribute of a variable.}
\usage{
GetTranslatedQuestionType(type, product.name)
}
\arguments{
\item{type}{A string indicating the required questiontype value.
Should be one of: "PickAny", "PickOne", "PickOneMulti", "Number",
"NumberMulti", "NumberGrid", "PickAnyGrid", "Text", "TextMulti",
"Ranking", or "Experiment".}

\item{product.name}{A string, which should be either Displayr or Q}
}
\description{
Helper function to use in Standard R items which require the user
to select variables of a certain question type / variable set
structure.
}
