% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{HasOutcome}
\alias{HasOutcome}
\title{Check if a formula contains an outcome variable}
\usage{
HasOutcome(formula)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}
}
\value{
logical
}
\description{
Check if a formula contains an outcome variable
}
