% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{HasSubset}
\alias{HasSubset}
\title{Check that a subset contains data}
\usage{
HasSubset(subset)
}
\arguments{
\item{subset}{The filter used to filter data in a model.}
}
\value{
true if the subset contains information
}
\description{
Check that a subset contains data
}
