% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IfThen}
\alias{IfThen}
\title{\code{IfThen}}
\usage{
IfThen(condition, TRUE.result, FALSE.result)
}
\arguments{
\item{condition}{A logical expression.}

\item{TRUE.result}{The object to be returned if \code{condition} is \code{TRUE}.}

\item{FALSE.result}{The object to be returned if \code{condition} is \code{FALSE}.}
}
\description{
If else statment as a single line.
}

