% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeat.R
\name{indicesToRetain}
\alias{indicesToRetain}
\title{indicesToRetain}
\usage{
indicesToRetain(names, at, length.x, ignore.case = TRUE, split = NULL)
}
\arguments{
\item{names}{The names of the elements.}

\item{at}{A \code{\link{vector}} or indicating which elements are to be removed.
When \code{at} is \code{\link{character}} elements with this name are removed. When \code{at}
is \code{\link{integer}}, elements with this position are removed. If \code{MARGIN} contains two
or more values, \code{at} may be a list, where each element of the list contains a vector.}

\item{length.x}{The length of the object that is to be subscripted.}

\item{ignore.case}{Ignores case if and when removing elements based on their names.}

\item{split}{Character delimiter to split \code{at} when \code{at} is a
\code{\link{character}}. Default is NULL which means that no splitting
occurs. See \code{\link{strsplit}}.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Worker function for \code{RemoveAt}.
}
