% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{InterceptExceptions}
\alias{InterceptExceptions}
\title{InterceptExceptions}
\usage{
InterceptExceptions(expr, warning.handler = NULL, error.handler = NULL)
}
\arguments{
\item{expr}{The expression whose warnings are to be intercepted.}

\item{warning.handler}{The function that handles intercepted warnings.}

\item{error.handler}{The function that handles intercepted errors.}
}
\value{
The value from evaluating \code{expr}.
}
\description{
This function intercepts warning and error messages produced
from running \code{expr} and passes them to \code{warning.handler} and
\code{error.handler}. Note that execution of expr is stopped after the
first error.
}
\examples{
addExclamationMark <- function(warn)
{
    warning(warn$message, "!")
}
## Intercepts a warning and rethrows it with an
## exclamation mark.
InterceptExceptions(warning("This is a warning"),
                    warning.handler = addExclamationMark)

addQuestionMark <- function(error)
{
    warning(error$message, "?")
}
## Intercepts the error from throwsAWarning and rethrows it with an
## exclamation mark.
InterceptExceptions(stop("Stop"), error.handler = addQuestionMark)
}
