% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{InterceptWarnings}
\alias{InterceptWarnings}
\title{InterceptWarnings}
\usage{
InterceptWarnings(expr, warning.handler)
}
\arguments{
\item{expr}{The expression whose warnings are to be intercepted.}

\item{warning.handler}{The function that handles intercepted warnings.}
}
\value{
The value of \code{expr}.
}
\description{
This function intercepts warning messages produced from running
\code{expr} and passes them to \code{warning.handler}.
}
