% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isrserver.R
\name{IsDevRServer}
\alias{IsDevRServer}
\title{IsDevRServer}
\usage{
IsDevRServer()
}
\value{
TRUE if running on the development R server. False otherwise.
}
\description{
This function indicates if it is being run on the development R
server.
}
