% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{IsQTable}
\alias{IsQTable}
\title{Identify Q Table objects}
\usage{
IsQTable(x)
}
\arguments{
\item{x}{Input to check if belonging to Q Table class or containing Q Table attributes}
}
\value{
Logical value \code{TRUE} if the input is identified as being a Q Table,
        \code{FALSE} otherwise
}
\description{
Inspects for the QTable class or checks if attributes exist
to identify legacy Q Table with the \code{"questions"} and
and \code{"name"} attributes
}
\note{
An attribute \code{"statistic"} is not guaranteed to be present, as
the names of the statistics computed may be present in the dimnames
}
