% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{LookupName}
\alias{LookupName}
\title{LookupName}
\usage{
LookupName(value, dictionary)
}
\arguments{
\item{value}{The value to be look uped in a \code{dictionary}.}

\item{dictionary}{A \code{\link{vector}} of values.}

\item{details}{Returns the name of the element in a vector corresponding to the \code{value}.
Where the vector is un-named, its index is
returned. Where the value appears multiple times a warning is provided.}
}
\description{
LookupName
}
