% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{OrderedToNumeric}
\alias{OrderedToNumeric}
\title{\code{OrderedToNumeric}}
\usage{
OrderedToNumeric(x)
}
\arguments{
\item{x}{An ordered factor.}
}
\description{
Convert an ordered factor to a numeric vector.
}

