% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{OutcomeName}
\alias{OutcomeName}
\title{Find the name of the outcome variable from a formula}
\usage{
OutcomeName(formula, data = NULL)
}
\arguments{
\item{formula}{A \code{\link{formula}} or a \code{\link[stats]{terms}} object.}

\item{data}{A \code{\link{data.frame}} containing the variables in the formula.
Currently, ignored.}
}
\value{
Character string giving the response variable name, or \code{NULL} if
no response is present in \code{formula}.
}
\description{
Find the name of the outcome variable from a formula
}
