% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{RemoveCharacterElements}
\alias{RemoveCharacterElements}
\title{RemoveCharacterElements}
\usage{
RemoveCharacterElements(names, names.to.remove = c("NET", "Total", "SUM"),
  split = "[;,]")
}
\arguments{
\item{names}{The \code{\link{vector}} of \code{\link{character}}..}

\item{names.to.remove}{A character vector or delimited string containing the
row labels to remove.}

\item{split}{Character delimiter to split \code{row.names.to.remove}
and \code{col.names.to.remove} on. Default is to split on either of
\code{","} or \code{";"}. Assumed to be a regular expression; see
\code{\link{strsplit}}.}
}
\value{
A vector containing the names that have not been removed.
}
\description{
Removes elements from a \code{\link{character}} vector, ignoring case.
}
\details{
Trailing spaces are removed and lower/upper case is ignored.
}
