% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeat.R
\name{removeFromDimension}
\alias{removeFromDimension}
\title{\code{RemoveAt}}
\usage{
removeFromDimension(
  x,
  at = NULL,
  MARGIN = 1L,
  ignore.case = TRUE,
  split = NULL
)
}
\arguments{
\item{x}{The object from which the elements are to be removed.}

\item{at}{A \code{\link{vector}} or indicating which elements are to be removed.
When \code{at} is \code{\link{character}} elements with this name are removed. When \code{at}
is \code{\link{integer}}, elements with this position are removed. If \code{MARGIN} contains two
or more values, \code{at} may be a list, where each element of the list contains a vector.}

\item{MARGIN}{An \code{\link{vector}} of integers giving the dimensions from which elements should be removed.}

\item{ignore.case}{Ignores case if and when removing elements based on their names.}

\item{split}{Character delimiter to split \code{at} when \code{at} is a
\code{\link{character}}. Default is NULL which means that no splitting
occurs. See \code{\link{strsplit}}.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Removes one or more elements from an object. Dimensions are never
             If all elements are removed from a vector, then a 0-length vector is returned.
}
