% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{Select.matrix}
\alias{Select.matrix}
\title{\code{Select}}
\usage{
\method{Select}{matrix}(x, from, to, MARGIN = NULL, drop = FALSE)
}
\arguments{
\item{x}{The object to be subscripted.}

\item{from}{The name of the first element to be selected.}

\item{to}{The name of the last element to be selected.}

\item{MARGIN}{The margin of the object to select from. Only used with arrays and matrices.}

\item{drop}{Whether to drop the final dimension. This parameter is only used for arrays and matrices.}
}
\description{
Subscripts a range of elements (e.g., from a vector or list), based on the
name of the elements.
}
