% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{Stop}
\alias{Stop}
\title{Error condition with special class to identify user errors}
\usage{
Stop(x, ...)
}
\arguments{
\item{x}{The message to pass to the user}

\item{...}{Additional arguments to pass to errorCondition}
}
\value{
An errorCondition with the "userError" class
}
\description{
Error condition with special class to identify user errors
}
