% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{StopUserError}
\alias{StopUserError}
\title{Error condition with special class to identify user errors}
\usage{
StopUserError(..., call. = FALSE, domain = NULL)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{call.}{Unused parameter, used to absorb usage from previous calls using stop}

\item{domain}{see \code{\link[base]{gettext}}.  If \code{NA}, messages will
    not be translated.}
}
\description{
Error condition with special class to identify user errors
}
