% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{StringIsFromControl}
\alias{StringIsFromControl}
\title{Specify that a string was generated by a Displayr control}
\usage{
StringIsFromControl(control.output)
}
\arguments{
\item{control.output}{A character vector to which the attribute
should be attached.}
}
\description{
Add an attribute to a character vector called is.control
which can tell other functions that a string originally
came from a combo box or listbox control.
}
