% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{UniquePlaceholders}
\alias{UniquePlaceholders}
\title{\code{UniquePlaceholders}}
\usage{
UniquePlaceholders(n.placeholders, n.characters = 64, padding = "")
}
\arguments{
\item{n.placeholders}{The length of the vector.}

\item{n.characters}{The number of characters in each entry.}

\item{padding}{Prefix and suffix characters for the placeholders.}
}
\value{
A vector of unique alphanumeric characters.
}
\description{
Generates a vector of unique alphanumeric characters, which can be used as
placeholders in text manipulation.
}
