% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AllIntegers}
\alias{AllIntegers}
\title{Check that all items in a variable are integers}
\usage{
AllIntegers(x)
}
\arguments{
\item{x}{The variable to check.}
}
\description{
Check that all items in a variable are integers
}

