% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{AnyNegative}
\alias{AnyNegative}
\title{Test whether a vector contains any negative values}
\usage{
AnyNegative(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
logical.
}
\description{
Test whether a vector contains any negative values
}
