% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{CopyAttributes}
\alias{CopyAttributes}
\title{\code{CopyAttributes}}
\usage{
CopyAttributes(data.with.attributes, data.without.attributes)
}
\arguments{
\item{data.with.attributes}{A \code{\link{data.frame}}.}

\item{data.without.attributes}{A \code{\link{data.frame}}.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Copies the "label" attribute for each for variable in a \code{\link{data.frame}}.
}

