% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{ExpectNoWarning}
\alias{ExpectNoWarning}
\title{\code{ExpectNoWarning}}
\usage{
ExpectNoWarning(code, string)
}
\arguments{
\item{code}{The code to evaluate for warnings.}

\item{string}{The string of text to be searched for in the warnings}
}
\description{
Checks to see if the printing of an object does not cause a warning to appear.
}
