% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeat.R
\name{RemoveAt}
\alias{RemoveAt}
\alias{RemoveAt.data.frame}
\alias{RemoveAt.array}
\alias{RemoveAt.matrix}
\alias{RemoveAt.ftable}
\title{\code{RemoveAt}}
\usage{
RemoveAt(x, at = NULL, MARGIN = NULL, ignore.case = TRUE, split = NULL)

\method{RemoveAt}{data.frame}(x, at = NULL, MARGIN = NULL, ignore.case = TRUE, split = NULL)

\method{RemoveAt}{array}(x, at = NULL, MARGIN = NULL, ignore.case = TRUE, split = NULL)

\method{RemoveAt}{matrix}(x, at = NULL, MARGIN = NULL, ignore.case = TRUE, split = NULL)

\method{RemoveAt}{ftable}(x, at = NULL, MARGIN = NULL, ignore.case = TRUE, split = NULL)
}
\arguments{
\item{x}{The object from which the elements are to be removed.}

\item{at}{A \code{\link{vector}} or indicating which elements are to be removed.
When \code{at} is \code{\link{character}} elements with this name are removed. When \code{at}
is \code{\link{integer}}, elements with this position are removed. If \code{MARGIN} contains two
or more values, \code{at} may be a list, where each element of the list contains a vector.}

\item{MARGIN}{An \code{\link{vector}} of integers giving the dimensions from which elements should be removed.}

\item{ignore.case}{Ignores case if and when removing elements based on their names.}

\item{split}{Character delimiter to split \code{at} when \code{at} is a
\code{\link{character}}. Default is NULL which means that no splitting
occurs. See \code{\link{strsplit}}.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Removes one or more elements from an object. Dimensions are never
             If all elements are removed from a vector, then a 0-length vector is returned.
}
\section{Methods (by class)}{
\itemize{
\item \code{RemoveAt(data.frame)}: Where only a \code{at} is provided, it is assumed
to be the variables of a data frame. Otherwise, the first \code{MARGIN}
is assumed the rows and the second the columns.

\item \code{RemoveAt(array)}: If \code{MARGIN} is not provided, the \code{at} argument
is used along all margins. If \code{at} is a list, then each
list element is sequentially used along each \code{MARGIN}
If \code{at} is not a list, then it all elements inside is
are removed across all margins.

\item \code{RemoveAt(matrix)}: Applies the array method

\item \code{RemoveAt(ftable)}: Applies the array method so that the dimnames of \code{ftable}

}}
