% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{StopUserError}
\alias{StopUserError}
\title{Error condition with special class to identify user errors}
\usage{
StopUserError(..., call. = FALSE)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{call.}{Unused parameter, used to absorb usage from previous calls using stop}
}
\description{
The inputs are not offered for translation, as they are assumed to be constructed using
\code{gettextf} or similar functions. That is, the inputs have already been offered for
translation before used as arguments.
}
